refactor : 오타 수정 및 Exception 위치 변경
