+++ seminar1/src/main/kotlin/com/wafflestudio/seminar/user/api/UserException.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.seminar.user.api.request
+package com.wafflestudio.seminar.user.api
 
 import org.springframework.http.HttpStatus
