+++ seminar1/src/main/kotlin/com/wafflestudio/seminar/user/api/UserExceptionHandler.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.seminar.user.api.request
+package com.wafflestudio.seminar.user.api
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
